refactor(nvim): use lazy.nvim `keys` for keymaps
Refactor plugin specifications to use the `keys` property for keymaps
instead of defining them inside `config` or `opts`. This makes the
configuration cleaner and leverages lazy-loading features.

For indent-blankline, use the config function to ensure it loads on
startup.

Also, defer the call to `SunglassesDisable` to a `VimEnter` autocommand
to prevent errors on startup due to lazy-loading.